+++ pushapkscript/src/mozapkpublisher/CHANGELOG.md
@@ -5,7 +5,19 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [upcoming]
 
 ### Changed
-* `push_apk` must be called as a function with parameters, rather than mocked `argv` arguments 
+* `push_apk` must be called as a function with parameters, rather than mocked `argv` arguments
+
+### Added
+* Options to skip specific sanity checks
+    * `--skip-check-ordered-version-codes`: Skip check that asserts version codes are different, x86 code > arm code
+    * `--skip-check-multiple-locales`: Skip check that asserts that apks all have multiple locales
+    * `--skip-check-same-locales`: Skip check that asserts that all apks have the same locales
+    * `--skip-checks-fennec`: Skip checks that are Fennec-specific (ini-checking, checking version-to-package-name compliance)
+    * `--skip-check-package-names`: Skip assertion that apks match a specified package name
+        * Note: if this isn't provided, then you must provide at least one `--expected-package-name`
+
+### Removed
+* Auto-detection of which package-name corresponds to which sanity checks 
 
 ## [0.13.0] - 2019-01-15